Added: -context specific decoders for all of the primitive times to go with the context specific encoders -unconfirmed/confirmed EventNotificationRequest structure/encoder/decoder -BACnetTimestamp structure/encoder/decoder -BACnetPropertyStates structure/encoder/decoder -BACnetDeviceObjectPropertyReference structure/encoder/decoder
Changes:
-decode_context_object_id::object_type changed to uint16_t
-explicit casts have been added to some functions to remove compiler warnings
-encode_bacnet_date::year behaviour has been changed slightly

+++ trunk/bacnet-stack/include/bacstr.h
@@ -79,7 +79,7 @@ extern "C" {
     uint8_t bitstring_bits_used(
         BACNET_BIT_STRING * bit_string);
 /* returns the number of bytes that a bit string is using */
-    int bitstring_bytes_used(
+    uint8_t bitstring_bytes_used(
         BACNET_BIT_STRING * bit_string);
     uint8_t bitstring_bits_capacity(
         BACNET_BIT_STRING * bit_string);
